backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
pickle-mixin==1.0.2
gunicorn==21.2.0
python-dateutil==2.8.2
joblib==1.3.2
pymongo==4.6.0
python-dotenv==1.0.0